fix(cli): move audit-licenses out of commands as it's not a command on its own
